rename : `Prompt.py` module renamed to `prompt.py`.
