+++ memor/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 """Memor modules."""
 from .params import MEMOR_VERSION
-from .Prompt import Prompt
+from .prompt import Prompt
 
 __version__ = MEMOR_VERSION
